途中経過

+++ Shogi3DBattle/header/Gold.h
@@ -5,7 +5,7 @@
 class Gold : public I_Piece
 {
 public:
-    unsigned short GetMovementBits() override; // 移動出来る範囲をビット列で返す
+    unsigned int GetMovementBits() override; // 移動出来る範囲をビット列で返す
 
     Gold(PlayerSide playerSide);
 };
