+++ Shogi3DBattle/header/Lance.h
@@ -5,7 +5,7 @@
 class Lance : public I_Piece
 {
 public:
-    unsigned short GetMovementBits() override; // 移動出来る範囲をビット列で返す
+    unsigned int GetMovementBits() override; // 移動出来る範囲をビット列で返す
 
     Lance(PlayerSide playerSide);
 };
