+++ Shogi3DBattle/header/Silver.h
@@ -5,7 +5,7 @@
 class Silver : public I_Piece
 {
 public:
-    unsigned short GetMovementBits() override; // 移動出来る範囲をビット列で返す
+    unsigned int GetMovementBits() override; // 移動出来る範囲をビット列で返す
 
     Silver(PlayerSide playerSide);
 };
